bot/requirements.txt
colorama==0.4.6
discord.py==2.3.1
pycryptodomex==3.18.0
pyfaceit==1.0.3
pystyle==2.9
pytz==2023.3
requests==2.28.2
ruamel.base==1.0.0
